Reorganize PDO related classes to be logically grouped.
