+++ src/FreeDSx/Ldap/Server/Backend/Storage/Adapter/Pdo/Connection/CoroutinePdoConnectionProvider.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace FreeDSx\Ldap\Server\Backend\Storage\Adapter\Pdo;
+namespace FreeDSx\Ldap\Server\Backend\Storage\Adapter\Pdo\Connection;
 
 use Closure;
 use FreeDSx\Ldap\Exception\RuntimeException;
